feat: 시큐리티 적용

+++ src/main/java/edu/ssafy/enjoytrip/controller/rest/InfoBoardRestController.java
@@ -3,7 +3,7 @@ package edu.ssafy.enjoytrip.controller.rest;
 import java.util.List;
 import java.util.Map;
 
-import javax.servlet.http.HttpSession;
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,21 +21,20 @@ import edu.ssafy.enjoytrip.dto.infoboard.InfoBoardCommentDto;
 import edu.ssafy.enjoytrip.dto.infoboard.InfoBoardCommentPageDto;
 import edu.ssafy.enjoytrip.dto.infoboard.InfoBoardDto;
 import edu.ssafy.enjoytrip.dto.infoboard.InfoBoardPageDto;
+import edu.ssafy.enjoytrip.jwt.JwtTokenProvider;
 import edu.ssafy.enjoytrip.service.infoboard.InfoBoardCommentService;
 import edu.ssafy.enjoytrip.service.infoboard.InfoBoardService;
 import edu.ssafy.enjoytrip.util.SizeConstant;
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequestMapping("/api/information")
+@RequiredArgsConstructor
 public class InfoBoardRestController {
 
-	private InfoBoardService infoBoardService;
-	private InfoBoardCommentService infoBoardCommentService;
-
-	public InfoBoardRestController(InfoBoardService infoBoardService, InfoBoardCommentService infoBoardCommentService) {
-		this.infoBoardService = infoBoardService;
-		this.infoBoardCommentService = infoBoardCommentService;
-	}
+	private final InfoBoardService infoBoardService;
+	private final InfoBoardCommentService infoBoardCommentService;
+	private final JwtTokenProvider tokenProvider;
 
 	// 공지사항 게시글 조회
 	@GetMapping("/list")
@@ -49,11 +48,11 @@ public class InfoBoardRestController {
 
 	// 게시글 작성
 	@PostMapping("/write")
-	public ResponseEntity<String> write(@RequestBody Map<String, String> map, HttpSession session) {
+	public ResponseEntity<String> write(@RequestBody Map<String, String> map, 
+			HttpServletRequest httpServletRequest) {
 		try {
-//			MemberDto memberDto = (MemberDto) session.getAttribute("userInfo");
-//			String userId = memberDto.getUserId();
-			String userId = "ssafy";
+			String token = httpServletRequest.getHeader("Authorization");
+			String userId = tokenProvider.getUserPk(token);
 			InfoBoardDto infoBoardDto = new InfoBoardDto();
 
 			infoBoardDto.setUserId(userId);
@@ -75,13 +74,21 @@ public class InfoBoardRestController {
 
 	// 게시글 수정
 	@PutMapping("/modify/{infoBoardId}")
-	public ResponseEntity<String> modify(@RequestBody Map<String, String> map, @PathVariable int infoBoardId) {
+	public ResponseEntity<String> modify(@RequestBody Map<String, String> map,
+			@PathVariable int infoBoardId,
+			HttpServletRequest httpServletRequest) {
 		try {
 			InfoBoardDto infoBoardDto = new InfoBoardDto();
 			infoBoardDto.setId(infoBoardId);
 			infoBoardDto.setTitle(map.get("title"));
 			infoBoardDto.setContent(map.get("content"));
-
+			
+			String token = httpServletRequest.getHeader("Authorization");
+			String userId = tokenProvider.getUserPk(token);
+			if (!userId.equals(infoBoardService.select(infoBoardId).getUserId())) {
+				throw new IllegalAccessException("잘못된 접근입니다.");
+			}
+			
 			infoBoardService.modify(infoBoardDto);
 			return new ResponseEntity<String>("공지사항 게시글 수정 성공", HttpStatus.OK);
 		} catch (Exception e) {
@@ -90,8 +97,15 @@ public class InfoBoardRestController {
 	}
 
 	@DeleteMapping("/delete/{infoBoardId}")
-	public ResponseEntity<String> delete(@PathVariable int infoBoardId) {
+	public ResponseEntity<String> delete(@PathVariable int infoBoardId,
+			HttpServletRequest httpServletRequest) {
 		try {
+			String token = httpServletRequest.getHeader("Authorization");
+			String userId = tokenProvider.getUserPk(token);
+			if (!userId.equals(infoBoardService.select(infoBoardId).getUserId())) {
+				throw new IllegalAccessException("잘못된 접근입니다.");
+			}
+			
 			infoBoardCommentService.deleteByPost(infoBoardId);
 			infoBoardService.delete(infoBoardId);
 			return new ResponseEntity<String>("공지사항 게시글 삭제 성공", HttpStatus.OK);
@@ -115,8 +129,6 @@ public class InfoBoardRestController {
 	public ResponseEntity<String> createComment(@PathVariable int infoBoardId, 
 			@RequestBody InfoBoardCommentDto commentDto){
 		try {
-			commentDto.setUserId("ssafy"); // 로그인 구현하게 되면 주석처리
-			commentDto.setInfoBoardId(infoBoardId); // 로그인 구현하게 되면 주석처리
 			infoBoardCommentService.write(commentDto);
 			return new ResponseEntity<String>("공지사항 게시글 댓글 작성 성공", HttpStatus.OK);
 		} catch (Exception e) {
@@ -128,6 +140,7 @@ public class InfoBoardRestController {
 	@PutMapping("/comment")
 	public ResponseEntity<String> updateComment(@RequestBody InfoBoardCommentDto commentDto){
 		try {
+			System.out.println(commentDto);
 			infoBoardCommentService.modify(commentDto);
 			return new ResponseEntity<String>("공지사항 게시글 댓글 수정 성공", HttpStatus.OK);
 		} catch (Exception e) {
